finish with advanced filters

+++ src/utils/utils.cpp
@@ -11,7 +11,7 @@ bool utils::is_integer(const std::string &s) {
     return std::regex_match(s, std::regex("[0-9]+"));
 }
 
-long utils::to_number(const std::string &s) {
+int utils::to_number(const std::string &s) {
     char *end;
     if(is_integer(s)) {
         return std::strtol(s.c_str(), &end, 10);
